+++ desktop/apps/auction2/components/client/AuctionApplication.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+export default class AuctionApplication extends Component {
+  render () {
+    return (
+      <div>
+        HI APP!
+      </div>
+    )
+  }
+}
