+++ desktop/apps/auction2/components/client/FilterSort.js
@@ -2,7 +2,7 @@ import _ from 'underscore'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
-import { updateSort } from 'desktop/apps/auction2/client/actions'
+import { updateSort } from 'desktop/apps/auction2/actions'
 
 function FilterSort (props) {
   const {
