+++ desktop/apps/auction2/routes.js
@@ -1,65 +1,69 @@
 import Articles from 'desktop/collections/articles.coffee'
-import ArticlesQuery from './queries/articles'
+import ArticlesQuery from 'desktop/apps/auction2/utils/queries/articles'
 import Auction from 'desktop/models/auction.coffee'
-import Index from './components/server/index'
-import MeQuery from './queries/me'
-import SaleQuery from './queries/sale'
-import footerItems from './footerItems'
+import Index from 'desktop/apps/auction2/components/server/index'
+import MeQuery from 'desktop/apps/auction2/utils/queries/me'
+import SaleQuery from 'desktop/apps/auction2/utils/queries/sale'
+import footerItems from 'desktop/apps/auction2/utils/footerItems'
 import metaphysics from 'lib/metaphysics.coffee'
 import get from 'lodash.get'
-import renderLayout from 'desktop/components/react/utils/render_layout'
+import renderReactLayout from 'desktop/components/react/utils/render_react_layout'
 import { getLiveAuctionUrl } from 'utils/domain/auctions/urls'
 
 export async function index (req, res, next) {
-  const { me } = await metaphysics({ query: MeQuery(req.params.id), req: req })
-  const { sale } = await metaphysics({ query: SaleQuery(req.params.id) })
-  const { articles } = await metaphysics({ query: ArticlesQuery(sale._id) })
-
-  res.locals.sd.AUCTION = sale
-
-  /**
-   * Ensure user model is in sync as updates to it during bidder registration
-   * route transitions will incidentally trigger a client-side reload.
-   * See: https://github.com/artsy/force/blob/master/desktop/lib/global_client_setup.coffee#L37
-   */
-  if (req.user) {
-    try {
-      const user = await req.user.fetch()
-      res.locals.sd.CURRENT_USER = {
-        ...res.locals.sd.CURRENT_USER,
-        ...user
-      }
-    } catch (error) {
-      console.log('(auction/routes.js) Error fetching user: ', error)
-    }
-  }
-
-  const auctionModel = new Auction(sale)
-  const auctionArticles = new Articles(articles)
-
   try {
-    const layout = renderLayout({
-      basePath: res.app.get('views'),
-      blocks: {
-        head: 'meta.jade',
-        body: Index
-      },
-      locals: {
-        ...res.locals,
-        articles: auctionArticles,
-        assetPackage: 'auctions',
-        auction: auctionModel,
-        bodyClass: 'body-header-fixed body-no-margins',
-        footerItems: footerItems,
-        isLiveOpen: auctionModel.isLiveOpen(),
-        liveAuctionUrl: getLiveAuctionUrl(auctionModel.get('id'), {
-          isLoggedIn: Boolean(me)
-        }),
-        me: me
-      }
-    })
+    const { me } = await metaphysics({ query: MeQuery(req.params.id), req: req })
+    const { sale } = await metaphysics({ query: SaleQuery(req.params.id) })
+    const { articles } = await metaphysics({ query: ArticlesQuery(sale._id) })
 
-    res.send(layout)
+    res.locals.sd.AUCTION = sale
+
+    /**
+     * Ensure user model is in sync as updates to it during bidder registration
+     * route transitions will incidentally trigger a client-side reload.
+     * See: https://github.com/artsy/force/blob/master/desktop/lib/global_client_setup.coffee#L37
+     */
+    if (req.user) {
+      try {
+        const user = await req.user.fetch()
+        res.locals.sd.CURRENT_USER = {
+          ...res.locals.sd.CURRENT_USER,
+          ...user
+        }
+      } catch (error) {
+        console.log('(auction/routes.js) Error fetching user: ', error)
+      }
+    }
+
+    const auctionModel = new Auction(sale)
+    const auctionArticles = new Articles(articles)
+
+    try {
+      const layout = renderReactLayout({
+        basePath: res.app.get('views'),
+        blocks: {
+          head: 'meta.jade',
+          body: Index
+        },
+        locals: {
+          ...res.locals,
+          articles: auctionArticles,
+          assetPackage: 'auctions',
+          auction: auctionModel,
+          bodyClass: 'body-header-fixed body-no-margins',
+          footerItems: footerItems,
+          isLiveOpen: auctionModel.isLiveOpen(),
+          liveAuctionUrl: getLiveAuctionUrl(auctionModel.get('id'), {
+            isLoggedIn: Boolean(me)
+          }),
+          me: me
+        }
+      })
+
+      res.send(layout)
+    } catch (error) {
+      next(error)
+    }
   } catch (error) {
     next(error)
   }
