+++ desktop/apps/auction2/test/reducers.js
@@ -1,8 +1,8 @@
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import auctions from 'desktop/apps/auction2/client/reducers'
-import * as actions from 'desktop/apps/auction2/client/actions'
-import { __RewireAPI__ as ActionsRewireApi } from 'desktop/apps/auction2/client/actions'
+import * as actions from 'desktop/apps/auction2/actions'
+import { __RewireAPI__ as ActionsRewireApi } from 'desktop/apps/auction2/actions'
 import sinon from 'sinon'
 
 describe('Reducers', () => {
