+++ desktop/apps/auction2/utils/footerItems.js
