+++ desktop/assets/auctions.coffee
@@ -18,7 +18,7 @@ routes =
   '''
   ^/auction2/[^/]+/?$
   /auction2/.*/confirm-registration
-  ''': require('../apps/auction2/client/index.js').default
+  ''': require('../apps/auction2/client.js').default
 
   '''
   /artist/.*/auction-results
